Extract community persistence
